requirements.txt
#
# Requirements for wger for production
#

Django>=1.9,<1.10
django-recaptcha
reportlab
dj-database-url==0.3.0
django_mobile
django-formtools>=1.0,<1.1
bleach
python-mimeparse
pillow
psycopg2
easy-thumbnails
django_compressor
icalendar
django-bootstrap-breadcrumbs
sphinx
django-sortedm2m
django-bower
invoke
requests
whitenoise==3.3.1
gunicorn==19.7.1


# REST API
djangorestframework>=3.2,<3.3
django-filter
django-tastypie
django-cors-headers

# Python3 compatibility
six
